db: change driver to mattn/go-sqlite3
Building now requires a C compiler. Setting it up in the actions was
fairly simple, though I haven't tested if the files run on Windows/Mac
respectively.

On the other hand, now bundling sqlite3 is not necessary, it is now
possible to simply use it as a shared library. Statically compiled
options are still in the build workflow, if wanted.